refactor: move address adapter to adapter namespace
