+++ src/Adapter/WcAddressAdapter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace MyParcelNL\WooCommerce\Factory;
+namespace MyParcelNL\WooCommerce\Adapter;
 
 use MyParcelNL\Pdk\Facade\Pdk;
 use WC_Cart;
